+++ Hospital System/src/main/java/Controllers/Admin/GetBestStrategies/GetBestByTime.java
@@ -7,7 +7,7 @@ public class GetBestByTime implements GetBestStrategy {
     StaffManaging sm = new StaffManager();
 
     @Override
-    public int findBest() {
+    public long findBest() {
         return sm.getBestStaffByTime();
     }
 }
